+++ src/scripts/singleMessage.js
@@ -5,7 +5,7 @@ const messageHtmlBuilder = {
         return `
         <div class="message-card"
         <section id="message-card-${singleMessage.id}">
-          <h2 id="message-header">Messages</h2>
+        <h2 id="message-header">Messages</h2>
           <p>${singleMessage.userId}:${singleMessage.message}</p>
           <button id="delete-message-${singleMessage.id}">Delete</button>
           <button id="edit-message-${singleMessage.id}">Edit</button>
